Clarified drivers section and document title
Updated the title to reflect that this document is also about interrupt-safety.
Added a link to the definition of "critical sections" (as used in this document)
Updated the section Synchronization levels to include a Venn diagram to help clarify the relationship between interrupt-safe and thread-safe
Updated the Drivers section to emphasis the need to first read the section "Synchronization levels". This is an attempt to reduce any confusion over the definition of thread-safe and interrupt-safe.