expand AnalogIn Description
The description is insufficient. Expanding for fuller explanation. 

Further removing Example 3 as it does nothing to add explanation or deeper understanding to the API. The code is not commented and is confusing. 

I would suggest pruning both the remaining examples (1 and 2) as they are likewise not terribly useful, but i'll leave them to be addressed at a later date.