Update Ticker and SleepManager based on feedback
Remove information about the 30 minute limitation from the Ticker
docs, as the newer API is not limited by this. Add a note to about the
higher power consumption of ticker. Add a note to SleepManager
indicating that sleep() should not need to be called normally. Also
explicitly reference classes which are safe to use in deep sleep mode.