Update windows installer instructions
Based on feedback during a workshop.

TLDR: the download link for the installer needs to be the first and biggest link in the section. We saw that during the workshop users just downloaded the first link they saw, which previously was just the mbed CLI link or the windows serial driver link. With this change we are making the first and foremost link the one for the installer to look similar to the OSX installer link.