Remove outdated references to FEATURE_NANOSTACK, COMMON_PAL or others.
In Mbed OS 5.10 all network related "features" have been removed.
This is transparent change to applications, as those code blocks are
now always enabled in the build.
Only visible change is that applications do not need to explicitly
request FEATURE_NANOSTACK or others, and targets do not need to enable
FEATURE_LWIP to provide ethernet driver.