workflow: define closure time
It has been known that more than 2 weeks for PR being idle means it can be closed.
We did not enforce it that often previously.

State this in the workflow, and reference it.

To be clear, closure does not mean "won't fix" but rather take your time and update
PR by reopening it.