workflow: add 3rd party updates as an exception
Updates like MbedTLS might contain more than just a bugfix. To keep it simple, they should be sent as one PR - version
update. Splitting this type of updates makes it difficult to maintain.

It should be stated what version we are getting and a reference to release notes (changelog could be - what this brings in).