Replace content in device-management.md (#1259)
* Replace content in device-management.md

Replace old content with links to the Device Management Client example for Mbed OS.

* Removing additional content

Ticket 5909 requested removing the section edited here (though I've left the first page with the new links) and the architecture section.
I'm still unclear about the Mbed CLI and Mbed IDE content - should we keep or discard. Following up with Will and Andy.

* Event example URL change

* Mbed CLI updates

Taking content from PR 1094 and adding to this PR, so all DM work is in one place. This doesn't answer the question of keep/discard, but it gives us a starting point for the discussion.

* Removing RawSerial

Unrelated to this, but breaking the build

* Removing Going to Production

* Removing Update from intro

* Update docs/ref-designs/pelion_ready/device-management.md

* Update docs/ref-designs/pelion_ready/device-management.md

* Auto stash before rebase of "origin/AnotherButler-patch-6"

Co-authored-by: Irit <irit.arkin@arm.com>