Update cli-reqs.md
In short: existing instructions for adding bash completion for mbed-cli didn't quite work.

I found that in a fresh install of Linux Mint 20 Cinnamon, I did not have a folder `~/.bash_completion.d`, and after adding one as per the Bash completion instructions in the mbed-os docs (and copying over the `mbed` script), the terminal still did not have bash completion for the mbed command. 

I found that I needed to add an instruction somewhere to make the terminal actually source that `mbed` script, since there was not anything in place that was looking in `~/.bash_completion.d`.

The actual method I used was adding:
`
for file in ~/.bash_completion.d/* ; do
    source "$file"
done
`
to my `~/.bashrc` file.