Mbed TLS no longer has a development-psa branch (#1361)
* PSA is in the main branch of Mbed TLS

The development-psa branch is an old feature branch that has long been
merged into development and master. Refer users to master, which is
the latest release.

Also remove the reference to a submodule. Mbed TLS no longer has any
submodules.

* Fix formatting glitch

* When building Mbed TLS, only build the library

Running the tests won't work on a cross build.

* In Mbed TLS, config.pl has been replaced by config.py

config.pl still exists but it only executes config.py. Change the
documentation to use config.py which is more future-proof and avoids
introducing a dependency on Perl.