porting: Rename MBED_BOOT_STACK_SIZE to MBED_CONF_TARGET_BOOT_STACK_SIZE
MBED_BOOT_STACK_SIZE is a duplicated definition of
MBED_CONF_TARGET_BOOT_STACK_SIZE, which we are removing as it is
causing issues in mbed-os.

This commit updates the porting guide to only mention
MBED_CONF_TARGET_BOOT_STACK_SIZE when referring to the define for the
boot stack size.